# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HighTac Local Platform API
  version: 2.0.0
  summary: Contract for the HighTac LAN management platform.
  description: |
    The Windows backend is the source of truth for stations, tags, products,
    bindings, commands, devices, and audit records. Timestamps are RFC 3339 UTC.

    Admin browser requests use the `hightac_session` cookie. Every mutating
    admin request also requires `X-CSRF-Token`. Approved Android installations
    use bearer identity and never receive MQTT credentials. The canonical retry
    key for retry-safe writes is the `Idempotency-Key` header; keys are scoped to
    the authenticated principal and retained for at least 24 hours.

    Approved Android devices also use bearer-authenticated, allowlisted mobile
    proxies for OpenAI-compatible Responses and JianDaoYun catalog reads. Third-
    party API credentials remain server-side and are never returned to clients.
  contact:
    name: HighTac Engineering
  license:
    name: Proprietary
...
servers:
  - url: http://{host}:{port}/api/v1
    description: HighTac platform on the trusted local network
    variables:
      host:
        default: 192.168.1.105
      port:
        default: "8088"
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      operationId: loginAdmin
      summary: Log in as the single site administrator.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              development:
                summary: Example administrator credentials; not a runtime account
                value:
                  username: example-admin
                  password: example-only-not-a-runtime-password
      responses:
        "200":
          description: Login succeeded. The session identifier is returned only as an HttpOnly cookie.
          headers:
            Set-Cookie:
              description: HttpOnly, SameSite=Strict admin session cookie.
              schema:
                type: string
                example: hightac_session=opaque; HttpOnly; SameSite=Strict; Path=/
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/logout:
    post:
      tags: [Authentication]
      operationId: logoutAdmin
      summary: Revoke the current admin session.
      security:
        - AdminSession: []
          CsrfToken: []
      responses:
        "204":
          description: Session revoked.
          headers:
            Set-Cookie:
              description: Expired admin session cookie.
              schema:
                type: string
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentAdmin
      summary: Return the current administrator.
      security:
        - AdminSession: []
      responses:
        "200":
          description: Current administrator.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/change-password:
    post:
      tags: [Authentication]
      operationId: changeAdminPassword
      summary: Change the current administrator password.
      security:
        - AdminSession: []
          CsrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed and other sessions revoked.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /light-commands:
    post:
      tags: [Commands]
      operationId: createLightCommand
      summary: Create a light-on or light-off command for a product or explicit tags.
      description: |
        Product targets resolve to all active bindings and are partitioned into
        packets of at most 20 tags. LIGHT_ON always uses beep, flashing, and a
        five-second duration. MQTT publish produces PUBLISHED, never CONFIRMED.
      security:
        - AdminSession: []
          CsrfToken: []
        - AndroidBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LightCommandCreate"
            examples:
              lightProductRed:
                summary: Light every actively bound tag for a product
                value:
                  action: LIGHT_ON
                  product_code: 1711A-ABA-PT
                  color: RED
              clearOneTag:
                summary: Clear one tag
                value:
                  action: LIGHT_OFF
                  tag_ids: [AD100000048F]
      responses:
        "202":
          description: Command accepted for asynchronous MQTT publication.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LightCommand"
              example:
                id: d7d74ed4-44d7-4934-af3a-a80d93aaf63f
                action: LIGHT_ON
                product_id: 12f2390e-301f-4c44-a90d-205b5ff0b39f
                product_code: 1711A-ABA-PT
                requested_color: RED
                status: ACCEPTED
                target_count: 1
                confirmed_count: 0
                unconfirmed_count: 0
                failed_count: 0
                created_at: "2026-07-16T08:15:30Z"
                published_at: null
                completed_at: null
                items: []
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /light-commands/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Commands]
      operationId: getLightCommand
      summary: Return command and per-tag acknowledgement state.
      security:
        - AdminSession: []
        - AndroidBearer: []
      responses:
        "200":
          description: Command state. PUBLISHED means sent, not executed.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LightCommand"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /stations/{station_id}/all-off:
    parameters:
      - $ref: "#/components/parameters/StationId"
    post:
      tags: [Commands]
      operationId: clearStation
      summary: Create a light-off command for every actively bound tag at the station.
      security:
        - AdminSession: []
          CsrfToken: []
        - AndroidBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StationAllOffRequest"
            example:
              confirmation: ALL OFF
      responses:
        "202":
          description: Station clear command accepted.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LightCommand"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: hightac_session
      description: Opaque HttpOnly SameSite=Strict administrator session cookie.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required together with AdminSession on every mutating browser request.
    AndroidBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-device-token
      description: Approved installation token stored in Android Keystore; only its hash is stored server-side.
    EnrollmentPollSecret:
      type: apiKey
      in: header
      name: X-Enrollment-Secret
      description: Short-lived installation-scoped secret returned once when enrollment is created.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUID. Reusing a key with the same method, path, actor,
        and normalized body returns the original result. Reusing it with a
        different request returns 409 IDEMPOTENCY_KEY_REUSED.
      schema:
        type: string
        format: uuid
      example: 2ec5fd45-0e4d-4be4-a6bd-3517118b64df
    StationId:
      name: station_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/StationId"
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
  headers:
    RequestId:
      description: Correlation identifier for logs and support diagnostics.
      schema:
        type: string
        format: uuid
    Location:
      description: Relative API path of the created or accepted resource.
      schema:
        type: string
        example: /api/v1/light-commands/d7d74ed4-44d7-4934-af3a-a80d93aaf63f
  responses:
...
    BadRequest:
      description: Request syntax or confirmation value is invalid.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: BAD_REQUEST
              message: The request could not be processed.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    Unauthorized:
      description: Authentication is missing, invalid, expired, or revoked.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: Valid authentication is required.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    Forbidden:
      description: The authenticated actor lacks permission, CSRF validation failed, or the Android device is not approved.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: FORBIDDEN
              message: This actor cannot perform the requested operation.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    NotFound:
      description: The requested resource does not exist.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: NOT_FOUND
              message: The requested resource was not found.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    Conflict:
      description: Current resource state conflicts with the request or an idempotency key was reused for a different request.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            activeBinding:
              value:
                error:
                  code: TAG_ALREADY_BOUND
                  message: The tag already has an active binding; explicit rebind is required.
                  details:
                    - field: tag_id
                      code: ACTIVE_BINDING_EXISTS
                      message: Tag AD100000048F is bound to product 1711A-OLD.
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
            idempotencyMismatch:
              value:
                error:
                  code: IDEMPOTENCY_KEY_REUSED
                  message: The idempotency key belongs to a different request.
                  details: []
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
            migrationPreviewMismatch:
              value:
                error:
                  code: MIGRATION_PREVIEW_MISMATCH
                  message: The preview token does not match this device and record set.
                  details: []
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
            migrationPreviewStale:
              value:
                error:
                  code: MIGRATION_PREVIEW_STALE
                  message: Authoritative binding state changed after the preview.
                  details: []
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
...
    ValidationError:
      description: One or more path, query, header, or body values failed validation.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: VALIDATION_ERROR
              message: One or more values are invalid.
              details:
                - field: tag_id
                  code: PATTERN_MISMATCH
                  message: Must match ^AD1[0-9A-F]{9}$.
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          description: Seconds before another attempt.
          schema:
            type: integer
            minimum: 1
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ServiceUnavailable:
      description: A required platform component such as the MQTT bridge or Windows service controller is unavailable.
      headers:
        Retry-After:
          description: Suggested retry delay in seconds.
          schema:
            type: integer
            minimum: 1
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: MQTT_UNAVAILABLE
              message: The broker is not available; no command was created.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
...
  schemas:
...
    Uuid:
      type: string
      format: uuid
      example: 2b16ae29-f7ab-4c44-9fc6-a0ca693d9c68
    Timestamp:
      type: string
      format: date-time
      example: "2026-07-16T08:15:30Z"
    StationId:
      type: string
      pattern: ^90A9F[0-9A-F]{7}$
      example: 90A9F1234567
    TagId:
      type: string
      pattern: ^AD1[0-9A-F]{9}$
      example: AD100000048F
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Error:
      type: object
      additionalProperties: false
      required: [code, message, details, request_id]
      properties:
        code:
          type: string
          pattern: ^[A-Z][A-Z0-9_]{1,63}$
        message:
          type: string
          minLength: 1
          maxLength: 1024
        details:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/ErrorDetail"
        request_id:
          $ref: "#/components/schemas/Uuid"
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: [string, "null"]
          maxLength: 256
        code:
          type: string
          pattern: ^[A-Z][A-Z0-9_]{1,63}$
        message:
          type: string
          minLength: 1
          maxLength: 1024
...
    AdminUser:
      type: object
      additionalProperties: false
      required: [id, username, must_change_password, is_active, created_at, last_login_at]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        username:
          type: string
          minLength: 1
          maxLength: 64
          example: Adam
        must_change_password:
          type: boolean
        is_active:
          type: boolean
        created_at:
          $ref: "#/components/schemas/Timestamp"
        last_login_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
        password:
          type: string
          minLength: 1
          maxLength: 1024
          writeOnly: true
    LoginResponse:
      type: object
      additionalProperties: false
      required: [user, csrf_token, session_expires_at]
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
        csrf_token:
          type: string
          minLength: 32
          maxLength: 256
          description: Non-cookie token sent as X-CSRF-Token on mutating admin requests.
        session_expires_at:
          $ref: "#/components/schemas/Timestamp"
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          minLength: 1
          maxLength: 1024
          writeOnly: true
        new_password:
          type: string
          minLength: 12
          maxLength: 1024
          writeOnly: true
...
    ProductCode:
      type: string
      minLength: 1
      maxLength: 128
      pattern: ^\S(?:.*\S)?$
      description: Trimmed and uppercased before uniqueness checks.
      example: 1711A-ABA-PT
...
    LightAction:
      type: string
      enum: [LIGHT_ON, LIGHT_OFF]
    LightColor:
      type: string
      enum: [RED, GREEN, BLUE, CYAN, PINK]
    CommandStatus:
      type: string
      enum: [ACCEPTED, PUBLISHED, CONFIRMED, PARTIALLY_CONFIRMED, UNCONFIRMED, FAILED, SUPERSEDED]
    CommandItemStatus:
      type: string
      enum: [PENDING, PUBLISHED, CONFIRMED, UNCONFIRMED, FAILED, SUPERSEDED]
    LightCommandCreate:
      type: object
      additionalProperties: false
      required: [action]
      properties:
        action:
          $ref: "#/components/schemas/LightAction"
        product_code:
          $ref: "#/components/schemas/ProductCode"
        tag_ids:
          type: array
          minItems: 1
          maxItems: 2000
          uniqueItems: true
          items:
            $ref: "#/components/schemas/TagId"
        color:
          $ref: "#/components/schemas/LightColor"
      oneOf:
        - required: [product_code]
          not:
            required: [tag_ids]
        - required: [tag_ids]
          not:
            required: [product_code]
      allOf:
        - if:
            properties:
              action:
                const: LIGHT_ON
            required: [action]
          then:
            required: [color]
        - if:
            properties:
              action:
                const: LIGHT_OFF
            required: [action]
          then:
            not:
              required: [color]
    CommandItem:
      type: object
      additionalProperties: false
      required:
        - id
        - tag_id
        - station_id
        - status
        - publish_attempts
        - published_at
        - confirmed_at
        - last_result_type
        - correlation
        - failure_code
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        tag_id:
          $ref: "#/components/schemas/TagId"
        station_id:
          $ref: "#/components/schemas/StationId"
        status:
          $ref: "#/components/schemas/CommandItemStatus"
        publish_attempts:
          type: integer
          minimum: 0
          maximum: 2
        published_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        confirmed_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        last_result_type:
          type: [integer, "null"]
          minimum: 0
          maximum: 255
        correlation:
          type: string
          enum: [NONE, HEURISTIC_STATION_TAG_TIME, HEURISTIC_STATE_MATCH]
          description: eStation acknowledgements contain no HighTac command ID.
        failure_code:
          type: [string, "null"]
          maxLength: 64
    LightCommand:
      type: object
      additionalProperties: false
      required:
        - id
        - action
        - product_id
        - product_code
        - requested_color
        - status
        - target_count
        - confirmed_count
        - unconfirmed_count
        - failed_count
        - created_at
        - published_at
        - completed_at
        - items
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        action:
          $ref: "#/components/schemas/LightAction"
        product_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        product_code:
          oneOf:
            - $ref: "#/components/schemas/ProductCode"
            - type: "null"
        requested_color:
          oneOf:
            - $ref: "#/components/schemas/LightColor"
            - type: "null"
        status:
          $ref: "#/components/schemas/CommandStatus"
        target_count:
          type: integer
          minimum: 0
        confirmed_count:
          type: integer
          minimum: 0
        unconfirmed_count:
          type: integer
          minimum: 0
        failed_count:
          type: integer
          minimum: 0
        created_at:
          $ref: "#/components/schemas/Timestamp"
        published_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        completed_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        items:
          type: array
          maxItems: 2000
          items:
            $ref: "#/components/schemas/CommandItem"
    StationAllOffRequest:
      type: object
      additionalProperties: false
      properties:
        confirmation:
          type: string
          const: ALL OFF
...